requirements.txt
# 2:22 DFIR Framework — Python Dependencies
# Backend + DFIR Core Engine

# ── Web Framework ───────────────────────────────────
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# ── Database ────────────────────────────────────────
SQLAlchemy>=2.0.30
PyMySQL>=1.1.0
mysql-connector-python>=8.3.0

# ── Authentication ──────────────────────────────────
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.1

# ── Validation & Serialization ──────────────────────
pydantic[email]>=2.7.0
python-dotenv>=1.0.0

# ── HTTP Client (for Paystack) ──────────────────────
httpx>=0.27.0

# ── PDF Report Generation ──────────────────────────
reportlab>=4.1.0
Pillow>=10.0.0

# ── DFIR Core: Semantic Analysis ────────────────────
numpy>=1.26.0
sentence-transformers>=2.6.0
scikit-learn>=1.4.0

# ── DFIR Core: LLM Narrative ───────────────────────
openai>=1.30.0

# ── Utilities ───────────────────────────────────────
tqdm>=4.66.0
rich>=13.0.0
